First step for invitations

+++ cz.fhsoft.poker.league/src/cz/fhsoft/poker/league/client/util/UUID.java
@@ -7,7 +7,7 @@ import com.google.gwt.user.client.Random;
 public class UUID {
 	
 	public static String generateUUID() {
-		long randomLong = (long) Random.nextInt() + 2*Integer.MAX_VALUE;
+		long randomLong = (long) Random.nextInt() + 2L * (long) Integer.MAX_VALUE;
 		long timeLong = new Date().getTime();
 		
 		return Long.toHexString(timeLong) + "-" + Long.toHexString(randomLong);
